+++ ffi/CHANGELOG.md
@@ -6,6 +6,7 @@
 - add player.setSpu to switch subtitle select
 - add player.spu to get current select spu track id
 - add player.audioTrack to get current select audio track id
+- add player.spuCount to get spu count
 
 ## 0.2.0